Establish layered desktop module boundaries (#105)
